feat: Enhance kernel scheduling with Event struct and related functionality
add `set` method for setting the event state to signaled and optionally unblocks any threads waiting on this event.

add `reset` method for  resetting the event state, indicating that the event is no longer in the signaled state.

add `poll` method for Checking the current state of the event without blocking, and optionally resets the state if autoReset is enabled.

add `wait` and `waitTimeout` methods for blocking the calling thread until the event is set(`wait`),
blocks the calling thread until the event is set or a specified timeout elapses(`waitTimeout`).

Add `waitMultiple` method to block the calling thread until any one of multiple events is set. This is currently a placeholder and not fully implemented (relies on the arch).